requirements.txt
tweepy
APScheduler
Flask
Pillow
python-dotenv